+++ src/content/worms/worms-management.js
@@ -1,12 +1,13 @@
 import { renderPage } from '../../helper/render-page.js'
 import escapeHTML from 'escape-html'
+import { renderNavigation } from './worms-basic.js'
 
 /**
  *
  * @param {import("../../data/types.js").App} App
  */
 export function setupWormsManagement(App) {
-  App.express.get('/worms/drafts', async (req, res) => {
+  App.express.get('/worms/your-bots', async (req, res) => {
     const user = req.user
     if (!user) {
       res.redirect('/')
@@ -20,9 +21,11 @@ export function setupWormsManagement(App) {
 
     renderPage(App, req, res, {
       page: 'worms-drafts',
-      heading: 'Deine Bots',
+      heading: 'Worms',
       backButton: false,
       content: `
+        ${renderNavigation(3)}
+
         ${bots
           .map(
             (bot) =>
@@ -36,17 +39,7 @@ export function setupWormsManagement(App) {
 
         <form action="/worms/drafts/create"><input name="name"> <input type="submit" value="Neuen Bot erstellen"></form>
 
-        <div style="height: 24px;"></div>
-
-        <h3>Testlauf</h3>
-
-        <form action="/worms/drafts/test-run">
-          <p>Rot: <select name="gId">${bots.map((bot) => `<option value="${bot.id}">${escapeHTML(bot.name)}</option>`)}</select></p>
-          <p>Grün: <select name="rId">${bots.map((bot) => `<option value="${bot.id}">${escapeHTML(bot.name)}</option>`)}</select></p>
-          <p><input type="submit" value="Starten"></p>
-        </form>
-
-        <div style="height: 200px;"></div>
+        <div style="height: 75px;"></div>
       `,
     })
   })
@@ -91,7 +84,7 @@ function think(dx, dy, board, x, y, dir, oppX, oppY) {
 `,
     })
 
-    res.redirect('/worms/drafts')
+    res.redirect('/worms/your-bots')
   })
 
   App.express.get('/worms/drafts/delete', async (req, res) => {
@@ -111,7 +104,7 @@ function think(dx, dy, board, x, y, dir, oppX, oppY) {
       where: { id, UserId: req.user.id },
     })
 
-    res.redirect('/worms/drafts')
+    res.redirect('/worms/your-bots')
   })
 
   App.express.get('/worms/drafts/edit', async (req, res) => {
@@ -140,10 +133,11 @@ function think(dx, dy, board, x, y, dir, oppX, oppY) {
       page: 'worms-drafts-edit',
       heading: 'Bot bearbeiten - ' + bot.name,
       backButton: false,
+      outsideOfContainer: true,
       content: `
         <p></p>
 
-        <p><button onClick="saveButtonClickedAndExit()">Speichern und Schließen</button><span style="display: inline-block; width: 30px;"></span><button onClick="saveButtonClicked()">Speichern</button><span style="display: inline-block; width: 30px;"></span>[<a href="/worms/drafts">schließen</a>]</p>
+        <p><button onClick="saveButtonClickedAndExit()">Speichern und Schließen</button><span style="display: inline-block; width: 30px;"></span><button onClick="saveButtonClicked()">Speichern</button><span style="display: inline-block; width: 30px;"></span>[<a href="/worms/your-bots">schließen</a>]</p>
 
         <p></p>
 
@@ -195,7 +189,7 @@ function think(dx, dy, board, x, y, dir, oppX, oppY) {
             const code = myEditor.getValue()
             const id = ${bot.id}
             postJson("/worms/drafts/save", {id, code}).then(() => {
-              window.location.href = '/worms/drafts'
+              window.location.href = '/worms/your-bots'
             })
           }
 
@@ -248,154 +242,5 @@ function think(dx, dy, board, x, y, dir, oppX, oppY) {
     bot.code = code
     await bot.save()
     res.send('ok')
-  })
-
-  App.express.get('/worms/drafts/test-run', async (req, res) => {
-    if (!req.user) {
-      res.redirect('/')
-      return
-    }
-
-    const rId = req.query.rId ? parseInt(req.query.rId.toString()) : NaN
-    const gId = req.query.gId ? parseInt(req.query.gId.toString()) : NaN
-
-    if (isNaN(rId) || isNaN(gId)) {
-      res.send('Missing gId or rId')
-      return
-    }
-
-    const rBot = await App.db.models.WormsBotDraft.findOne({
-      where: { id: rId, UserId: req.user.id },
-    })
-    const gBot = await App.db.models.WormsBotDraft.findOne({
-      where: { id: gId, UserId: req.user.id },
-    })
-
-    if (!rBot || !gBot) {
-      res.send('Bot not found')
-      return
-    }
-
-    renderPage(App, req, res, {
-      page: 'worms-test-run',
-      heading: 'Testlauf',
-      backButton: false,
-      content: `
-        <p><a href="/worms/drafts">zurück</a></p>
-
-        <p style="text-align: center;">Rot: ${rBot.name} / Grün: ${gBot.name}</p>
-
-        <script src="/worms/wormer.js"></script>
-
-        <script
-          src="/worms/quickjs.js"
-          type="text/javascript"
-        ></script>
-
-        <div id="board"></div>
-        <p style="text-align: right;"><input type="checkbox" onClick="wormer.toggleTurbo()"/> Turbo</p>
-
-        <div style="height: 200px;"></div>
-
-        <script>
-
-          QJS.getQuickJS().then((QuickJS) => {
-            const runtimeRed = QuickJS.newRuntime()
-            runtimeRed.setMemoryLimit(1024 * 640)
-            runtimeRed.setMaxStackSize(1024 * 320)
-            let cyclesRed = { val: 0 }
-            runtimeRed.setInterruptHandler(() => {
-              cyclesRed.val++
-            })
-            const ctxRed = runtimeRed.newContext()
-
-            // ---------------------------------------
-            const logHandle = ctxRed.newFunction("log", (...args) => {
-              const nativeArgs = args.map(ctxRed.dump)
-              console.log("QuickJS:", ...nativeArgs)
-            })
-            const consoleHandle = ctxRed.newObject()
-            ctxRed.setProp(consoleHandle, "log", logHandle)
-            ctxRed.setProp(ctxRed.global, "console", consoleHandle)
-            consoleHandle.dispose()
-            logHandle.dispose()
-            // -------------------------
-
-            try {
-              ctxRed.evalCode(\`${rBot.code.replace(/`/g, '\\`')}\`)
-            } catch {}
-
-            const runtimeGreen = QuickJS.newRuntime()
-            runtimeGreen.setMemoryLimit(1024 * 640)
-            runtimeGreen.setMaxStackSize(1024 * 320)
-
-            let cyclesGreen = { val: 0 }
-            runtimeGreen.setInterruptHandler(() => {
-              cyclesGreen.val++
-            })
-            const ctxGreen = runtimeGreen.newContext()
-            try {
-              ctxGreen.evalCode(\`${gBot.code.replace(/`/g, '\\`')}\`)
-            } catch {}
-
-            // ---------------------------------------
-            const logHandle2 = ctxGreen.newFunction("log", (...args) => {
-              const nativeArgs = args.map(ctxGreen.dump)
-              console.log("QuickJS:", ...nativeArgs)
-            })
-            const consoleHandle2 = ctxGreen.newObject()
-            ctxGreen.setProp(consoleHandle2, "log", logHandle2)
-            ctxGreen.setProp(ctxGreen.global, "console", consoleHandle2)
-            consoleHandle2.dispose()
-            logHandle2.dispose()
-            // -------------------------
-
-
-
-            const red = (() => {
-              return (dx, dy, board, x, y, dir, oppX, oppY) => {
-                const callScriptRed = \`
-                  think(74, 42, \${JSON.stringify(board)}, \${x}, \${y}, \${dir}, \${oppX}, \${oppY});
-                \`
-                let newDirRed = -1
-                try {
-                  cyclesRed.val = 0
-                  const resultRed = ctxRed.unwrapResult(ctxRed.evalCode(callScriptRed))
-                  newDirRed = ctxRed.getNumber(resultRed)
-                  resultRed.dispose()
-                  console.log('red cycles (10k)', cyclesRed.val)
-                } catch(e) {
-                  alert('Fehler in Rot: ' + e) 
-                }
-                return newDirRed
-              }
-            })()
-            const green = (() => {
-              return (dx, dy, board, x, y, dir, oppX, oppY) => {
-                const callScriptGreen = \`
-                  think(74, 42, \${JSON.stringify(board)}, \${x}, \${y}, \${dir}, \${oppX}, \${oppY});
-                \`
-                let newDirGreen = -1
-                try {
-                  cyclesGreen.val = 0
-                  const resultGreen = ctxGreen.unwrapResult(ctxGreen.evalCode(callScriptGreen))
-                  newDirGreen = ctxGreen.getNumber(resultGreen)
-                  resultGreen.dispose()
-                  console.log('green cycles (10k)', cyclesGreen.val)
-                } catch(e) {
-                  alert('Fehler in Grün: ' + e) 
-                }
-                return newDirGreen
-              }
-            })()
-            const wormer = new Wormer(document.getElementById('board'), red, green)
-            window.wormer = wormer
-            wormer.run()
-          })
-
-          
-        </script>
-      `,
-    })
   })
 }
